backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.18
pypdf==3.17.0
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.0
azure-identity==1.15.0
requests==2.31.0

